chore: move header to components/docs
